unify core spec
